requirements.txt
edge-srt-to-speech
openai
python-dotenv